renamed directives.js into focus.js and removed unused mock in test
