+++ shell/plugins/focus.js
