+++ shell/plugins/index.js
@@ -4,7 +4,7 @@
  * These are included in a function that can be explictly called, so that we can be sure
  * of the execution order, rather than importing them at the top of a file.
  */
-import focusDirective from '@shell/plugins/directives';
+import focusDirective from 'plugins/focus';
 import cleanHtmlDirective from 'plugins/clean-html-directive';
 import vCleanTooltip from 'plugins/clean-tooltip-directive';
 import positiveIntNumber from '@shell/plugins/positive-int-number.js';
